% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.tw.r
\name{cut.tw}
\alias{cut.tw}
\title{Cut a TWINSPAN tree into groups of data}
\usage{
\method{cut}{tw}(x, level = NULL, clusters = NULL, long.output = F, ...)
}
\arguments{
\item{x}{Object of the class \code{'tw'} created by the function \code{twinspan}.}

\item{level}{The hierarchical level at which the tree should be cut (default is NULL - the level will be set up according to maximum number of levels created by \code{twinspan}). Works for both standard and modified TWINSPAN, but differs in behaviour (see Details).}

\item{clusters}{Number of resulting clusters (only for results of modified TWINSPAN).}

\item{long.output}{Logical - should the output contain only vector with group assignment (FALSE), or also additional information (TRUE)?}

\item{...}{Arguments passed to default \code{cut} function. Not implemented here.}
}
\value{
If \code{long.output = FALSE} returns a vector of the same length as number of samples with assignement of sample into clusters. 
If \code{long.output = TRUE}, output is a \code{data.frame} with three columns: 
\itemize{
\item \code{plot.no} with the original plot number, 
\item \code{group} with the assignment to cluster, 
\item \code{class} with the binary string created by \code{twinspan.exe} defining the hiearchical relationships of individual samples.}
}
\description{
Cut a TWINSPAN tree into groups of data
}
\details{
Argument \code{level} can be used for results of both standard and modified TWINSPAN, while argument \code{clusters} is applicable only on results of modified TWINSPAN. In the case of modified TWINSPAN the behavior of \code{level} and \code{clusters} differ - the first indicates at which hierarchical level the resulting tree is cut, while the second says how many final groups should be recognized. Since in modified TWINSPAN there is one division at each hierarchical level, the relationship between \code{level} and \code{clusters} is \code{level + 1 = clusters}.

 In case of standard TWINSPAN, number of resulting clusters depends not only on hierarchical level of cutting, but also on other stopping criteria (mainly \code{min.group.size}). Number of groups =< 2^\code{level}.
}
\author{
David Zeleny (zeleny.david@gmail.com)
}

