% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cutree.tw.r
\name{cutree.tw}
\alias{cutree.tw}
\title{Cut a TWINSPAN tree into groups of data}
\usage{
cutree.tw(tw, level = NULL, long.output = F)
}
\arguments{
\item{tw}{object created by the function \code{twinspan}}

\item{level}{the hierarchical level at which the tree should be cut (default is NULL - the level will be set up according to maximum number of levels created by \code{twinspan})}

\item{long.output}{logical - should the output contain only vector with group assignment (FALSE), or also additional information (TRUE)?}
}
\value{
If \code{long.output = FALSE} returns a vector of the same length as number of samples with assignement of sample into clusters.
If \code{long.output = TRUE}, output is a \code{data.frame} with three columns:
\itemize{
\item \code{plot.no} with the original plot number,
\item \code{group} with the assignment to cluster,
\item \code{class} with the binary string created by \code{twinspan.exe} defining the hiearchical relationships of individual samples.}
}
\description{
Cut a TWINSPAN tree into groups of data
}
\author{
David Zeleny (zeleny.david@gmail.com)
}

