% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.tw.r
\name{plot.tw}
\alias{plot.tw}
\title{Draws TWINSPAN dendrogram}
\usage{
\method{plot}{tw}(tw, level = NULL)
}
\arguments{
\item{tw}{object created by the function \code{twinspan}}

\item{level}{xxx}
}
\description{
Draws TWINSPAN dendrogram
}
\author{
David Zeleny (zeleny.david@gmail.com)
}

