% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.tw.r
\name{print.tw}
\alias{print.tw}
\title{Printing the results of TWINSPAN}
\usage{
\method{print}{tw}(x, what = NULL, clusters = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{'tw'}, created by \code{twinspan} function.}

\item{what}{What should be printed. Default = NULL (only standard output, i.e. \code{data.frame} with three columns stored in \code{x$classif} is returned). Should be (partial match to) one of  \code{c('species.names', 'order.plots', 'reading.data', 'input.parameters', 'levels', 'table', 'twi')}. See Details for explanation.}

\item{clusters}{Vector of integers. For which clusters (in case of modified TWINSPAN) the results should be returned? Default = NULL (i.e. information will be returned for all created clusters).}

\item{...}{Arguments passed to default \code{print} function. Not implemented here.}
}
\value{
If \code{what} is \code{'species.names'} or \code{'order.plots'}, output is a \code{data.frame}. If \code{what} is \code{'reading.data'}, \code{'input.parameters'}, \code{'levels'}, \code{'table'} or \code{'twi'}, output is printed into standard connection, usually console (if not redirected by \code{sinc}). This output can be recorded using \code{capture.output} as a character vector (each element equals to one line of original console output). Such captured output could be e.g. written to file using \code{write.table} function (see Examples).

In case of modified TWINSPAN, the results for individual hierarchical divisions are separated by lines of '@@@@@...'.
}
\description{
Generic \code{print} function applicable on object of class \code{tw}, created by function \code{twinspan}. With appropriate setting of its arguments returns detail results of TWINSPAN (including indicator species or two-way sorted table).
}
\details{
Function \code{print.tw} can access different outputs generated by TWINSPAN program, namely those stored in \code{tw.PUN} and \code{tw.TWI} files in \code{/exec} subdirectory. Printed output somewhat differ for standard and modified TWINSPAN, since modified TWINSPAN records information separately for each division, while standard TWINSPAN generate only one output files for all divisions.

Argument \code{what} can have the following values:
\itemize{
\item \code{species.names} Prints \code{data.frame} with the full species names (columns \code{full.name}) and 8-letter abbreviations generated for TWINSPAN (and used in output files) by function \code{make.cepnames} from \code{vegan} (column \code{abbrev.name}).
\item \code{order.plots} Prints \code{data.frame} with plots sorted by TWINSPAN algorithm (column \code{order} indicates the sequential order of the plots, column \code{plot.no} stores the original plot number).
\item \code{reading.data} Prints the header of TWINSPAN output, containing information about which data have be read by the algorithm
\item \code{input.parameters} Prints the second section of TWINSPAN output with setting of input parameters.
\item \code{levels} Prints the third section of TWINSPAN output with results (e.g. indicator species) for particular levels of division.
\item \code{table} Prints the two-way sorted table of plots and species (the fourth section of TWINSPAN output)
\item \code{twi} Prints all four sections of TWINSPAN output (\code{reading.data}, \code{input.parameters}, \code{levels} and \code{table}).}
}
\examples{
data (danube)
tw <- twinspan (danube$spe)
print (tw, 'table')

## The output of print function could be captured by "capture.output" function,
## and writen into a file using e.g. "write.table":
\dontrun{
write.table (file = 'table.txt', capture.output (print (tw, 'table')), quote = F, row.names = F)
system (shell ('notepad.exe table.txt'))}
}
\seealso{
\code{\link{summary.tw}}, \code{\link{twinspan}}
}

