% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/twinspan-package.r
\docType{package}
\name{twinspan-package}
\alias{twinspan-package}
\title{TWo-way INdicator SPecies ANalaysis (R wrapper around twinspan.exe)}
\description{
TWo-way INdicator SPecies ANalaysis (R wrapper around twinspan.exe)
}
\author{
David Zeleny
}

