% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/twinspan.r
\name{twinspan}
\alias{summary.tw}
\alias{twinspan}
\title{(Modified) TWINSPAN in R}
\usage{
twinspan(com, modif = F, cut.levels = c(0, 2, 5, 10, 20),
  min.group.size = 5, levels = 6, clusters = 5, diss = "bray",
  min.diss = NULL, mean.median = "mean", show.output.on.console = FALSE,
  quiet = TRUE, ...)

\method{summary}{tw}(object, ...)
}
\arguments{
\item{com}{Community data (\code{data.frame} or \code{matrix}).}

\item{modif}{Should the modified TWINSPAN algorithm be used? (logical, value, default = FALSE, i.e. standard TWINSPAN)}

\item{cut.levels}{Pseudospecies cut levels (default = \code{c(0,2,5,10,20)}). Should not exceed 9 cut levels.}

\item{min.group.size}{Minimum size of the group, which should not be further divided (default = 5).}

\item{levels}{Number of hierarchical levels of divisions (default = 6, should be between 0 and 15). Applies only for standard TWINSPAN (\code{modif = FALSE}).}

\item{clusters}{Number of clusters generated by modified TWINSPAN (default = 5). Applies only for modified TWINSPAN (\code{modif = TRUE}).}

\item{diss}{Dissimilarity used to calculate cluster heterogeneity (default = 'bray'); see Details for more information. Applies only for modified TWINSPAN (\code{modif = TRUE}).}

\item{min.diss}{Minimum dissimilarity under which the cluster will not be divided further (default = NULL, which means that the stopping rule is based on number of clusters (parameter \code{clusters})). Currently not implemented.}

\item{mean.median}{Should be the average dissimilarity of cluster calculated as mean or median of all between sample dissimilarities within the cluster? (default = \code{'mean'}, alternative is \code{'median'})}

\item{show.output.on.console}{Logical; should the function communicating with \code{twinspan.exe} show the output (rather long) of TWINSPAN program on console? Default = \code{FALSE}. Argument passsed via function \code{shell} into \code{system}.}

\item{quiet}{Logical; should the function reading TWINSPAN output files (tw.TWI and tw.PUN) be quiet and not report on console number of items it has read? Default = \code{TRUE}, means the function is quiet. Argument passed into function \code{scan}.}

\item{...}{Other (rarely used) TWINSPAN parameters passed into function \code{\link{create.tw.dat}}}

\item{object}{Object of the class \code{'tw'}.}
}
\value{
\code{twinspan} returns object of the class \code{'tw'}, which is a list with the following items:
\itemize{
\item \code{classif} data frame with three columns: \code{order} - sequential number of plot, \code{plot.no} - original number of plot (\code{row.names} in community matrix \code{com}), \code{class} - binomial code with hieararchical result of TWINSPAN classification.
\item \code{twi} vector (if \code{modif = FALSE}) or list (if \code{modif = TRUE}) of complete machine-readable output of TWINSPAN algorithm read from *.TWI file. In case of modified TWINSPAN (\code{modif = TRUE}) it is a list with number of items equals to number of clusters.
\item \code{spnames} data frame with two columns: \code{full.name} - full scientific species name (\code{names (com)}), \code{abbrev.name} - eight-digits abbreviation created by \code{make.cepnames} function from \code{vegan}.
\item \code{modif} logical; was the result calculated using standard TWINSPAN (\code{modif = FALSE}) or its modified version (\code{modif = TRUE})?
}
}
\description{
Calculates TWINSPAN (TWo-way INdicator SPecies ANalaysis, Hill 1979) and its modified version according to Rolecek et al. (2009)
}
\details{
XXXXX

Arguments \code{show.output.on.console} and \code{quiet} regulates how 'verbal' will be \code{twinspan} function while running. Default setting (\code{show.output.on.console = FALSE, quiet = TRUE}) supress all the output. Setting \code{quiet = FALSE} has only minor effect - it reports how many items have been read in each step of analysis from the output files (tw.TWI and tw.PUN). In contrary setting \code{show.output.on.console = TRUE} prints complete output generated by \code{twinspan.exe} program on console. Argument \code{show.output.on.console} has similar behavior as the argument of the same name in function \code{\link{system}}, but the value is not directly passed to this function. Output could be captured using function \code{capture.output} - see examples below.
}
\examples{
## Modified TWINSPAN on traditional Ellenberg's Danube meadow dataset, projected on DCA
## and compared with original classification into three vegetation types made by tabular sorting:
library (twinspanR)
library (vegan)
data (danube)
res <- twinspan (danube$spe, modif = TRUE, clusters = 4)
k <- cut (res)
dca <- decorana (danube$spe)
par (mfrow = c(1,2))
ordiplot (dca, type = 'n', display = 'si', main = 'Modified TWINSPAN')
points (dca, col = k)
for (i in c(1,2,4)) ordihull (dca, groups = k, show.group = i, col = i,
 draw = 'polygon', label = TRUE)
ordiplot (dca, type = 'n', display = 'si', main = 'Original assignment\\n (Ellenberg 1954)')
points (dca, col = danube$env$veg.type)
for (i in c(1:3)) ordihull (dca, groups = danube$env$veg.type,
 show.group = unique (danube$env$veg.type)[i], col = i,
 draw = 'polygon', label = TRUE)

## To capture the console output of twinspan.exe into R object, use the following:
\dontrun{
out <- capture.output (tw <- twinspan (danube$spe, show.output.on.console = T))
summary (tw)           # returns summary of twinspan algorithm
cat (out, sep = '\\n')  # prints the captured output
write.table (out, file = 'out.txt', quot = F, row.names = F) # writes output to 'out.txt' file
}
}
\author{
Mark O. Hill wrote the original Fortran code, which has been adopted by Petr Smilauer into dynamic library twindll.dll and executable program \code{twinspan.exe}; the R function maintaining communication between R and \code{twinspan.exe} was written by David Zeleny (zeleny.david@gmail.com).
}
\references{
\itemize{
\item Hill M.O. (1979): TWINSPAN - A FORTRAN program for arranging multivariate data in an ordered two-way table by classification of the individuals and attributes. Section of Ecology and Systematics, Cornel University, Ithaca, New York.
\item Rolecek J., Tichy L., Zeleny D. & Chytry M. (2009): Modified TWINSPAN classification in which the hierarchy respects cluster heterogeneity. Journal of Vegetation Science, 20: 596-602.
}
}
\seealso{
\code{\link{cut.tw}}, \code{\link{create.tw.dat}}
}

