% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinspanR-package.r
\docType{package}
\name{twinspanR}
\alias{twinspanR}
\title{TWo-way INdicator SPecies ANalaysis (R wrapper around twinspan.exe)}
\description{
Calculates TWINSPAN (classical and modified version)
}
\author{
David Zeleny
}
