% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.tw.r
\name{print.tw}
\alias{print.tw}
\title{Printing the results of TWINSPAN
Generic \code{print} function applicable on object of class \code{tw}, created by function \code{twinspan}. With appropriate setting of its arguments returns detail results of TWINSPAN (including indicator species or two-way sorted table).}
\usage{
\method{print}{tw}(x, what = NULL, clusters = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{'tw'}, created by \code{twinspan} function.}

\item{what}{Vector of character strings, indicating which part of TWINSPAN output should be printed. Default = NULL (only standard output, i.e. \code{data.frame} with three columns stored in \code{x$classif} is returned). Should be one (or more) of  \code{c('species.names', 'reading.data', 'input.parameters', 'levels', 'table', 'twi')}.}

\item{clusters}{Vector of integers. For which clusters (in case of modified TWINSPAN) the results should be returned? Default = NULL (i.e. information will be returned for all created clusters).}

\item{...}{Arguments passed to default \code{print} function. Not implemented here.}
}
\description{
Printing the results of TWINSPAN
Generic \code{print} function applicable on object of class \code{tw}, created by function \code{twinspan}. With appropriate setting of its arguments returns detail results of TWINSPAN (including indicator species or two-way sorted table).
}

