% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twinspanR-package.r
\docType{package}
\name{twinspanR}
\alias{twinspanR}
\alias{twinspanR-package}
\title{TWo-way INdicator SPecies ANalaysis (R wrapper around twinspan.exe)}
\description{
TWo-way INdicator SPecies ANalaysis (R wrapper around twinspan.exe)
}
\author{
David Zeleny
}

